Revert "Revert "Move `VideoCallClient` into its own crate (#142)" (#147)"
This reverts commit 672a7321e40007701ea693881ec16a8bc9b813f5.
