+++ Cargo.toml
@@ -5,5 +5,6 @@ members = [
     "bot",
     "src-tauri",
     "types",
+    "videocall-client",
     "yew-ui"
 ]
