+++ videocall-client/Cargo.toml
@@ -1,0 +1,92 @@
+[package]
+name = "videocall-client"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+aes = "0.8.3"
+anyhow = "1"
+cbc = { version = "0.1.2", features = ["alloc"] }
+gloo = "0.8.0"
+gloo-timers = "0.2.6"
+gloo-utils = "0.1"
+js-sys = "0.3"
+log = "0.4.19"
+protobuf = "3.2.0"
+rand = { version = "0.8.5", features = ["std_rng", "small_rng"] }
+rsa = "0.9.2"
+types = { path= "../types"}
+wasm-bindgen = "0.2.78"
+wasm-bindgen-futures = "0.4.30"
+yew = { version = "0.20" }
+yew-websocket = "1.0.1"
+yew-webtransport = "0.3.2"
+
+[dependencies.web-sys]
+version = "0.3.60"
+features = [
+    "AudioTrack",
+    "AudioData",
+    "AudioEncoder",
+    "AudioEncoderInit",
+    "AudioEncoderConfig",
+    "AudioDecoder",
+    "AudioDecoderInit",
+    "AudioDecoderConfig",
+    "AudioContext",
+    "BaseAudioContext",
+    "GainOptions",
+    "GainNode",
+    "console",
+    "CodecState",
+    "CanvasRenderingContext2d",
+    "EncodedAudioChunk",
+    "EncodedAudioChunkInit",
+    "EncodedAudioChunkType",
+    "EncodedVideoChunk",
+    "EncodedVideoChunkInit",
+    "EncodedVideoChunkType",
+    "MediaStreamAudioDestinationNode",
+    "AudioDestinationNode",
+    "AudioContextOptions",
+    "AudioDataCopyToOptions",
+    "HtmlCanvasElement",
+    "HtmlImageElement",
+    "Navigator",
+    "MediaDevices",
+    "MediaStream",
+    "MediaStreamTrack",
+    "MediaTrackSettings",
+    "MediaStreamTrackProcessor",
+    "MediaStreamTrackProcessorInit",
+    "MediaStreamTrackGenerator",
+    "MediaStreamTrackGeneratorInit",
+    "WritableStream",
+    "WritableStreamDefaultWriter",
+    "MediaStreamAudioSourceNode",
+    "HtmlVideoElement",
+    "MediaStreamConstraints",
+    "ReadableStream",
+    "ReadableStreamGetReaderOptions",
+    "ReadableStreamDefaultReader",
+    "VideoEncoder",
+    "VideoEncoderInit",
+    "VideoEncoderConfig",
+    "VideoEncoderEncodeOptions",
+    "VideoFrame",
+    "VideoTrack",
+    "VideoDecoder",
+    "VideoDecoderConfig",
+    "VideoDecoderInit",
+    "LatencyMode",
+    "HtmlAudioElement",
+    "AudioDataInit",
+    "AudioSampleFormat",
+    "TransformStream",
+    "MediaDeviceInfo",
+    "MediaDeviceKind",
+    "MediaTrackConstraints",
+    "CanvasRenderingContext2d"
+]
