+++ videocall-client/README.tpl
@@ -1,0 +1,25 @@
+<!---  Do not manually edit README.md!   The file is created automatically using 'cargo readme'.
+
+    Changes should be made in:
+
+        * src/lib.rs
+        * README.tpl
+
+    Rebuilding:
+
+        $ cd videocall-client
+        $ cargo readme > README.md
+
+--->
+
+# {{crate}}
+
+Version: {{version}}
+
+---
+
+{{readme}}
+
+---
+
+*Readme created using [cargo-readme](https://github.com/webern/cargo-readme) using [README.tpl](./README.tpl)*
