+++ videocall-client/rust-toolchain.toml
@@ -1,0 +1,3 @@
+[toolchain]
+channel = "nightly-2023-01-27"
+targets = [ "wasm32-unknown-unknown" ]
