+++ videocall-client/src/connection/mod.rs
@@ -1,3 +1,4 @@
+#[allow(clippy::module_inception)]
 mod connection;
 mod task;
 mod webmedia;
