+++ videocall-client/src/media_devices/mod.rs
@@ -2,4 +2,4 @@ mod media_device_access;
 mod media_device_list;
 
 pub use media_device_access::MediaDeviceAccess;
-pub use media_device_list::MediaDeviceList;
+pub use media_device_list::{MediaDeviceList, SelectableDevices};
