--- yew-ui/src/model/mod.rs
@@ -1,8 +1,0 @@
-pub mod client;
-pub mod connection;
-pub mod decode;
-pub mod encode;
-pub mod media_devices;
-pub mod wrappers;
-
-pub use client::{VideoCallClient, VideoCallClientOptions};
